test(integration): split EndpointSmokeTests into per-endpoint files (#430 cleanup)
The Phase 5 EndpointSmokeTests bundle covered every read-only endpoint
in one file (24 tests across File / Container / Folder / Ecosystem GETs
plus a few mutating ones). That made the test layout asymmetric vs the
new MutatingEndpointTests / BootstrapEndpointTests, and bundled the
File-related tests with the Folder-related tests with the Ecosystem-
related ones — hard to find tests for a specific handler.

Split into per-handler files matching the *Endpoints.cs source layout:

- FileEndpointTests       — CheckFileInfo, GetFile, ecosystem_pointer,
                            ancestry (8 tests, 200/404/412 paths)
- ContainerEndpointTests  — CheckContainerInfo, ecosystem_pointer,
                            ancestry, children + extension filter
                            (8 tests, 200/404 paths)
- FolderEndpointTests     — CheckFolderInfo, children (4 tests)
- EcosystemEndpointTests  — CheckEcosystem, root_container_pointer
                            (2 tests)

Share a single backend via the new ReadOnlyEndpointsFixture +
[CollectionDefinition("ReadOnlyEndpoints")] — read-only tests don't
mutate the FS root, so the WopiBackend boots once and the four classes
each take an [Collection]-injected reference. The fixture also owns
the MintFileTokenAsync / MintContainerTokenAsync helpers (previously
duplicated per-class).

The three Lock cycle / mismatch / unknown-override tests that lived in
EndpointSmokeTests are mutating in nature — moved into
MutatingEndpointTests as:
- Lock_Unlock_Cycle_Roundtrips (multi-step LOCK→GET_LOCK→REFRESH→UNLOCK
  state survival across HTTP requests)
- Lock_OnLockedDifferentId_Returns_409_WithExistingLock (renamed from
  Lock_Mismatch — distinct from Lock_OnLockedSameId_Refreshes which
  exercises the refresh branch)
- Unknown_Override_Returns_404 (WopiOverrideMatcherPolicy fallthrough)

Net: same 96 tests pass + 3 OIDC skips. Coverage on the read-only
endpoint files unchanged (still 100%).
